refactor(buildinfo): use a macro instead of a build script
Use one of our own macros to read the board name from a shell
environment variable we provide from the build system.
This was previously done using a build script, leveraging the fact that
cargo exposed `--cfg` flags as environment variables, but using a macro
is simpler, faster, and safer.