fixup! threads: PoC multicore support for rp2040
Only schedule a context switch if thread allocation (to a core) changed.

TODO:
Schedule the context switch on the core that is affected. In most cases,
this is the currently running cores, but not always (e.g. when on
another core a low priority thread is running that should be preempted).