ci(hax): update the hax setup
The `Makefile` is copied from from the following Gist:
https://gist.github.com/W95Psp/4c304132a1f85c5af4e4959dd6b356c3
This Gist is recommend by the official hax setup documentation:
https://hacspec.org/book/quick_start/intro.html#setup-the-crate-you-want-to-verify

The Makefile is modified to not try to download `hacl-star`, as it is
already present thanks to the nix flake in the hax Action.

This still only does lax typechecking, and does *not* prove panic
freedom of the runqueue.