fix(threads/smp): sched on other core if sched is already pending
When a thread becomes ready, the scheduler on the core with the lowest
prio running thread will be triggered. If now another thread becomes
ready as well and the scheduler didn't have a chance to run yet (e.g.
because interrupts are still disabled), the same scheduler will be
triggered again, but only one thread is then selected and can run.
The other thread is "skipped".
To solve this, the scheduler on the other core should be triggered as
well in this scenario so that both schedulers get the most recent state
and the two highest prio threads are run.