feat(device-id/nrf): Use traits to enforce consistency, offer general accessor
This makes the defmt dependency of riot-rs-embassy-common unconditional
to allow enfocing it in a trait.

The "convenience" wrapper riot_rs::identity::device_identity is added
because a simple single method call would otherwise not only require
users to dig into ::arch, but also to import a trait form -common (which
they may not even depend on).