refactor(coap): Alter CoAP output strategy
Previously, the coapcore printed to a provided logger. That is
inconvenient as a general logging strategy in a mostly defmt
environment, and moreover, it needs passing writers across layers which
they should not pass. That output is ignored now.

Instead (to keep the interactive output interesting), the CoAP client
demo parts now report to CoAP live output.