fix(coap): update Lakers
This commit contains the minimal changes to make things work, and does
not yet reap support for non-int C_x.

This breaks compatibility with aiocoap when credentials are sent by
value, as is the case in some examples. This is part of the fallout of a
previous Lakers bug [303], and due to be fixed in an upcoming aiocoap
update.

[303]: https://github.com/openwsn-berkeley/lakers/issues/303